Temp++
Animation int형에서 float 바꿔서
EffectMgr에 int 형으로 되어 있는 Speed를 float 형으로 바꿈, 이미지 추가

+++ Pixel-master/Space/EffectMgr.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "EffectMgr.h"
 
-EffectMgr::EffectMgr(std::wstring filename, int start, int speed, int end, Vec2 Pos)
+EffectMgr::EffectMgr(std::wstring filename, int start, float speed, int end, Vec2 Pos)
 {
 	Effect = new Animation();
 	Effect->AddContinueFrame(filename, start, end);
